fix(ts_codegen): route Option* union arms via capnp_to_ts_type; drop dead Text/Data checks
Review findings #2 and #3 on PR #471.

#2 (correctness): `union_variant_data_type` sent `Option*` wrapper struct types
(e.g. `OptionUint32`) down the struct-pointer branch, emitting `OptionUint32 | null`
instead of `number | undefined`. Option* wrappers now short-circuit through
`capnp_to_ts_type`, which encodes absence as `undefined` and must not get the
struct-pointer `| null` suffix. Latent today (no such union schema yet) but fixed.

#3 (nit): removed the dead `|| tn == "Text" || tn == "Data"` conditions —
`is_primitive` (is_primitive_capnp_type) already covers Text and Data.

Also moved `union_variant_data_type` from `interfaces.rs` to `mod.rs` as
`pub(crate)` so the parser-result type aliases (finding #1) can share it.
